removed some of the k6 checkout flow into separate k6 requests and moved headers to their own file.
